Test alignment against button glitch
button glitch:

When having a button in a function that changes the values of $boxParameter1 and $boxParameter2 and having function2 that uses the same button to change the value of 3rd parameter without touching $boxParameter1 and $boxParameter2, a glitch happens.

The glitch: the button in function2 effects the $boxParameter1 and $boxParameter2 the same as the same button effected them in the last function used.

This only happens because function2 does not work with  $boxParameter1 and $boxParameter2.